chore: refactor core factories structure
