--- packages/core/core/src/factories/constants.ts
@@ -1,3 +1,0 @@
-export const symbols = {
-  CustomController: Symbol('StrapiCustomCoreController'),
-} as const;
