--- packages/core/core/src/factories/index.ts
@@ -1,3 +1,0 @@
-// Defines the public API for factories
-
-export { createCoreController, createCoreService, createCoreRouter } from './factories';
